+++ src/shapes/IText/ITextKeyBehavior.ts
@@ -101,18 +101,6 @@ export abstract class ITextKeyBehavior<
       'compositionend',
       this.onCompositionEnd.bind(this)
     );
-
-    if (!this._clickHandlerInitialized && this.canvas) {
-      this.canvas.upperCanvasEl.addEventListener(
-        'click',
-        this.onClick.bind(this)
-      );
-      this._clickHandlerInitialized = true;
-    }
-  }
-
-  onClick() {
-    this.hiddenTextarea && this.hiddenTextarea.focus();
   }
 
   /**